+++ wrapper/asset/load.js
@@ -1,10 +1,12 @@
-const loadPost = require('../request/post_body');
-const sessions = require('../data/sessions');
-const asset = require('./main');
+/***
+ * asset load route
+ */
+const loadPost = require("../request/post_body");
+const asset = require("./main");
 
 module.exports = function (req, res, url) {
 	switch (req.method) {
-		case 'GET': {
+		case "GET": {
 			const match = req.url.match(/\/assets\/([^/]+)$/);
 			if (!match) return;
 
@@ -15,20 +17,17 @@ module.exports = function (req, res, url) {
 			return true;
 		}
 
-		case 'POST': {
-			if (url.path != '/goapi/getAsset/' && url.path != '/goapi/getAssetEx/') return;
+		case "POST": {
+			if (url.path != "/goapi/getAsset/" && url.path != "/goapi/getAssetEx/") return;
 			loadPost(req, res).then(data => {
-				const mId = data.movieId || data.presaveId || sessions.get(req);
 				const aId = data.assetId || data.enc_asset_id;
 
-				const b = asset.loadLocal(mId, aId);
-				sessions.set({ movieId: mId }, req);
+				const b = asset.load(aId);
 				if (b) {
-					res.setHeader('Content-Length', b.length);
-					res.setHeader('Content-Type', 'audio/mp3');
+					res.setHeader("Content-Length", b.length);
+					res.setHeader("Content-Type", "audio/mp3");
 					res.end(b);
-				}
-				else {
+				} else {
 					res.statusCode = 404;
 					res.end();
 				};