+++ changelog.md
@@ -17,6 +17,8 @@ This changelog is based off [Keep a Changelog](https://keepachangelog.com) and ~
   - Remade character uploading
   - Removed build numbering
   - Removed the .sh files in the wrapper folder
+  - Removed Google Wavenet voices
+  - Removed IBM Watson voices
   - Node.js installer is now 17.6.0
 ### Fixed
   - Polly voices